==> notebooks/1. Downloading data.ipynb <==
from urllib.request import urlretrieve
from pyspark.sql import SparkSession
import pandas as pd
from pyspark.sql import functions as F
from pyspark.sql import SparkSession
from pyspark.sql.functions import *

spark = (
 SparkSession.builder.appName("project 1")
 .config("spark.sql.repl.eagerEval.enabled", True)
 .config("spark.executor.memory","4G")
 .config("spark.driver.memory","2G")
 .config("spark.sql.parquet.cacheMetadata", "true")
 .config("spark.sql.session.timeZone", "Etc/UTC")
 .config('spark.driver.maxResultSize', '2048m')
 .getOrCreate()
)

import os

# from the current `tute_1` directory, go back two levels to the `MAST30034` directory

# check if it exists as it makedir will raise an error if it does exist
if not os.path.exists('../data/'):
    os.makedirs('../data/')

if not os.path.exists('../data/raw'):
    os.makedirs('../data/raw')

if not os.path.exists('../data/curated'):
    os.makedirs('../data/curated')

# Downloading data in Jan 2023 for comparison #

output_dir = "../data/raw/2023-01.parquet"
urlretrieve("https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2023-01.parquet", output_dir)

output_dir = "../data/raw/green_tripdata_2023-01.parquet"
urlretrieve("https://d37ci6vzurychx.cloudfront.net/trip-data/green_tripdata_2023-01.parquet", output_dir)

output_dir = "../data/raw/fhv_tripdata_2023-01.parquet"
urlretrieve("https://d37ci6vzurychx.cloudfront.net/trip-data/fhv_tripdata_2023-01.parquet", output_dir)

output_dir = "../data/raw/fhvhv_tripdata_2023-01.parquet"
urlretrieve("https://d37ci6vzurychx.cloudfront.net/trip-data/fhvhv_tripdata_2023-01.parquet", output_dir)

# Downloading other six months data #

output_dir = "../data/raw/2022-12.parquet"
urlretrieve("https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2022-12.parquet", output_dir)

output_dir = "../data/raw/2023-02.parquet"
urlretrieve("https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2023-02.parquet", output_dir)

output_dir = "../data/raw/2023-03.parquet"
urlretrieve("https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2023-03.parquet", output_dir)

output_dir = "../data/raw/2023-04.parquet"
urlretrieve("https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2023-04.parquet", output_dir)

output_dir = "../data/raw/2023-05.parquet"
urlretrieve("https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2023-05.parquet", output_dir)

# Transforming data type for each file #

sdf_feb = spark.read.parquet("../data/raw/2023-02.parquet")

consistent_col_casing = [F.col(col_name).alias(col_name.lower()) for col_name in sdf_feb.columns]
sdf_feb = sdf_feb.select(*consistent_col_casing)
sdf_schema = sdf_feb.schema


sdf_malformed = spark.read.parquet('../data/raw/2022-12.parquet')

sdf_malformed = sdf_malformed \
    .select([F.col(c).cast(sdf_schema[i].dataType) for i, c in enumerate(sdf_malformed.columns)])

sdf_malformed \
    .coalesce(1) \
    .write \
    .mode('overwrite') \
    .parquet('../data/curated/2022-12')

sdf_malformed = spark.read.parquet('../data/raw/2023-01.parquet')

sdf_malformed = sdf_malformed \
    .select([F.col(c).cast(sdf_schema[i].dataType) for i, c in enumerate(sdf_malformed.columns)])

sdf_malformed \
    .coalesce(1) \
    .write \
    .mode('overwrite') \
    .parquet('../data/curated/2023-01')

sdf_malformed = spark.read.parquet('../data/raw/2023-02.parquet')

sdf_malformed = sdf_malformed \
    .select([F.col(c).cast(sdf_schema[i].dataType) for i, c in enumerate(sdf_malformed.columns)])

sdf_malformed \
    .coalesce(1) \
    .write \
    .mode('overwrite') \
    .parquet('../data/curated/2023-02')

sdf_malformed = spark.read.parquet('../data/raw/2023-03.parquet')

sdf_malformed = sdf_malformed \
    .select([F.col(c).cast(sdf_schema[i].dataType) for i, c in enumerate(sdf_malformed.columns)])

sdf_malformed \
    .coalesce(1) \
    .write \
    .mode('overwrite') \
    .parquet('../data/curated/2023-03')

sdf_malformed = spark.read.parquet('../data/raw/2023-04.parquet')

sdf_malformed = sdf_malformed \
    .select([F.col(c).cast(sdf_schema[i].dataType) for i, c in enumerate(sdf_malformed.columns)])

sdf_malformed \
    .coalesce(1) \
    .write \
    .mode('overwrite') \
    .parquet('../data/curated/2023-04')

sdf_malformed = spark.read.parquet('../data/raw/2023-05.parquet')

sdf_malformed = sdf_malformed \
    .select([F.col(c).cast(sdf_schema[i].dataType) for i, c in enumerate(sdf_malformed.columns)])

sdf_malformed \
    .coalesce(1) \
    .write \
    .mode('overwrite') \
    .parquet('../data/curated/2023-05')
