==> notebooks/2. Preprocessing.ipynb <==
from pyspark.sql import SparkSession
import pandas as pd
from pyspark.sql import functions as F
from pyspark.sql import SparkSession
from pyspark.sql.functions import *
import numpy as np
from urllib.request import urlretrieve

# Exploring Dataset #

spark = (
 SparkSession.builder.appName("project 1")
 .config("spark.sql.repl.eagerEval.enabled", True)
 .config("spark.executor.memory","6G")
 .config("spark.driver.memory","6G")
 .config("spark.sql.parquet.cacheMetadata", "true")
 .config("spark.sql.session.timeZone", "Etc/UTC")
 .config('spark.driver.maxResultSize', '2048m')
 .getOrCreate()
)

Yellow_Taxi_Jan = spark.read.parquet('../data/raw/2023-01.parquet')
Green_Taxi_Jan = spark.read.parquet('../data/raw/green_tripdata_2023-01.parquet')
FHV_Jan = spark.read.parquet('../data/raw/fhv_tripdata_2023-01.parquet')
FHV_HV_Jan = spark.read.parquet('../data/raw/fhvhv_tripdata_2023-01.parquet')

print("Number of trips recorded by Yellow Taxi: ", Yellow_Taxi_Jan.count())
print("Number of trips recorded by Green Taxi: ", Green_Taxi_Jan.count())
print("Number of trips recorded by FHV: ", FHV_Jan.count())
print("Number of trips recorded by FHV_hv: ", FHV_HV_Jan.count())

## Feature Selection and Data Cleaning ##

#Adding new features by transforming the exist features
def new_feature(sdf):
    return sdf.withColumn('datetime', date_format('tpep_pickup_datetime', 'yyyy-MM-dd'))\
        .withColumn('duration', unix_timestamp('tpep_dropoff_datetime') - unix_timestamp('tpep_pickup_datetime'))\
        .withColumn('hourly_earn', F.col('total_amount')/(unix_timestamp('tpep_dropoff_datetime') - unix_timestamp('tpep_pickup_datetime'))*3600)\
        .withColumn('Month', date_format('tpep_pickup_datetime', 'MM').cast('int'))\
        .withColumn('Hour', date_format('tpep_pickup_datetime', 'HH').cast('int'))\
        .withColumn('Weekend', dayofweek('tpep_pickup_datetime').isin(1,7))\
        .withColumn('Airport', (F.col('airport_fee') > 0).cast('BOOLEAN'))\
        .withColumn('Congestion',(F.col('congestion_surcharge') > 0).cast('BOOLEAN'))

#Deleting unusual instance
def feature_clean(sdf):
    return sdf.where((F.col('duration') > 60) & (F.col('duration') < 3600*5))\
        .where((F.col('trip_distance') < 280) & (F.col('trip_distance') > 0))\
        .where(F.col('hourly_earn') > 0)\
        .where(F.col('total_amount') > 0)\
        .where(F.col('congestion_surcharge') >= 0)\
        .where(F.col('passenger_count') >= 0)\
        .where(F.col('payment_type') == 1)\
        .dropna()
        

sdf = spark.read.parquet('../data/curated/*')

print('Num of raw data :', sdf.count())
print('Null data ratio: ', sdf.dropna().count()/sdf.count())
#Derive the percentage of instances without null value

sdf = new_feature(sdf)
sdf = feature_clean(sdf)

print('Num of data after being filtered: ', sdf.count())

## Merge New Dataset ##

weather_dataset = spark.read.csv("../data/newyork_weather.csv", header = True)

weather_dataset

merged_sdf = sdf.join(weather_dataset, on = 'datetime', how = 'leftouter')

# Transforming them to float datatype from string
merged_sdf = merged_sdf.withColumn("temp", F.col("temp").cast('float'))
merged_sdf = merged_sdf.withColumn("precip", F.col("precip").cast('float'))
merged_sdf = merged_sdf.withColumn("windspeedmean", F.col("windspeedmean").cast('float'))
merged_sdf = merged_sdf.withColumn("visibility", F.col("visibility").cast('float'))

## Export dataset to parquet file ##

merged_sdf.write.mode('overwrite').parquet('../data/curated/merged_sdf.parquet')

# 1% instances from the original dataset as sample
df = merged_sdf.sample(0.01, seed=0).toPandas()
df.to_parquet('../data/curated/sample_data.parquet')
