==> notebooks/5. F_test.ipynb <==
import pandas as pd
from pyspark.sql import functions as F
from pyspark.sql import SparkSession
from pyspark.sql.functions import *
from statsmodels.formula.api import ols, glm
import statsmodels.api as sm

df = pd.read_parquet('../data/curated/sample_data_remove_outliers.parquet')

#Since there is tip_amount = 0, log transformation would have problem with it
log_transformed_data = np.log(df.loc[df['tip_amount'] > 0,'tip_amount'])

filtered_df = df[df['tip_amount'] != 0]
log_transformed_data = np.log(filtered_df['tip_amount'])

#Transforming those numeric datatype to categorical to prepare f test
filtered_df['PULocationID'] = pd.Categorical(filtered_df['PULocationID'])
filtered_df['DOLocationID'] = pd.Categorical(filtered_df['DOLocationID'])

## ANOVA test ##

# Create an ANOVA model
model = ols('log_transformed_data ~ Weekend+Airport+Congestion+PULocationID+DOLocationID', data=filtered_df).fit()

# Perform the ANOVA analysis
anova_table = sm.stats.anova_lm(model, typ=2)

print(anova_table)
